# REFACTORING - unpack/packMsg
## Change description
- `unpack` function do not unpack a message but parse, so changed the function name
- Move `unpack` function location to reduce compilation time (Actually, `unpack` is not an event handled by the handler but just a action. `packMsg` is same reason)
